fix: resolve all CI failures — explicit ruff config, auto-fix lint errors, fix conditional React hooks, gate build on CI
- Add explicit [tool.ruff.lint] config with select/ignore rules so CI
  is deterministic (self-hosted runner had global config selecting broad
  rules that differed from local defaults)
- Auto-fix 190+ ruff errors (import sorting, comprehensions, format)
- Manually fix 28 remaining errors (exception chaining, logging,
  implicit Optional, self-assignments, list comprehensions)
- Fix 2 conditional useMemo hooks in SeriesTimeline.tsx (moved hooks
  above early return to comply with Rules of Hooks)
- Gate Build and Push workflow on CI passing via workflow_run trigger
  (previously build ran independently and shipped broken images)

+++ src/doc_intelligence_hub/modules/statements/detector.py
@@ -329,7 +329,7 @@ def _build_diagnostic_entry(
 
 def _build_pattern(frequency: str, dates: list[date], grace_period_days: int) -> AnalysisPattern:
     days = [value.day for value in dates]
-    variance = int(round(statistics.pstdev(days))) if len(days) > 1 else 0
+    variance = round(statistics.pstdev(days)) if len(days) > 1 else 0
     interval_days = [(dates[index + 1] - dates[index]).days for index in range(len(dates) - 1)]
     interval_variance = statistics.pstdev(interval_days) if len(interval_days) > 1 else 0
     confidence = max(0.55, min(0.99, 1.0 - (interval_variance / 10)))
